requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
python-docx>=1.1.0
pyyaml>=6.0
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0